third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-pause/debugger-pause-on-exception-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that pause on exception works.

Script source was shown.
Script execution paused.
Call stack:
    0) throwAnException (debugger-pause-on-exception.js:12)
    1) handleClick (debugger-pause-on-exception.js:17)
    [setTimeout]
    0)  (debugger-pause-on-exception.js:31)
Paused on exceptionReferenceError: unknown_var is not defined
Script execution resumed.
